+++ SCrawler/MainFrame.Designer.vb
@@ -29,6 +29,7 @@ Partial Public Class MainFrame : Inherits System.Windows.Forms.Form
         Dim MENU_VIEW_SEP_2 As System.Windows.Forms.ToolStripSeparator
         Dim TRAY_SEP_1 As System.Windows.Forms.ToolStripSeparator
         Dim MENU_VIEW_SEP_4 As System.Windows.Forms.ToolStripSeparator
+        Dim MENU_DOWN_ALL_SEP_1 As System.Windows.Forms.ToolStripSeparator
         Dim resources As System.ComponentModel.ComponentResourceManager = New System.ComponentModel.ComponentResourceManager(GetType(MainFrame))
         Me.MENU_SETTINGS = New System.Windows.Forms.ToolStripDropDownButton()
         Me.BTT_SETTINGS = New System.Windows.Forms.ToolStripMenuItem()
@@ -41,7 +42,11 @@ Partial Public Class MainFrame : Inherits System.Windows.Forms.Form
         Me.BTT_CHANNELS = New System.Windows.Forms.ToolStripButton()
         Me.BTT_DOWN_SAVED = New System.Windows.Forms.ToolStripButton()
         Me.BTT_DOWN_SELECTED = New System.Windows.Forms.ToolStripButton()
-        Me.BTT_DOWN_ALL = New System.Windows.Forms.ToolStripButton()
+        Me.MENU_DOWN_ALL = New System.Windows.Forms.ToolStripDropDownButton()
+        Me.BTT_DOWN_ALL = New System.Windows.Forms.ToolStripMenuItem()
+        Me.BTT_DOWN_SITE = New System.Windows.Forms.ToolStripMenuItem()
+        Me.BTT_DOWN_ALL_FULL = New System.Windows.Forms.ToolStripMenuItem()
+        Me.BTT_DOWN_SITE_FULL = New System.Windows.Forms.ToolStripMenuItem()
         Me.BTT_DOWN_VIDEO = New System.Windows.Forms.ToolStripButton()
         Me.BTT_DOWN_STOP = New System.Windows.Forms.ToolStripButton()
         Me.MENU_VIEW = New System.Windows.Forms.ToolStripDropDownButton()
@@ -74,6 +79,7 @@ Partial Public Class MainFrame : Inherits System.Windows.Forms.Form
         Me.USER_CONTEXT = New System.Windows.Forms.ContextMenuStrip(Me.components)
         Me.BTT_CONTEXT_DOWN = New System.Windows.Forms.ToolStripMenuItem()
         Me.BTT_CONTEXT_DOWN_LIMITED = New System.Windows.Forms.ToolStripMenuItem()
+        Me.BTT_CONTEXT_DOWN_DATE_LIMIT = New System.Windows.Forms.ToolStripMenuItem()
         Me.BTT_CONTEXT_EDIT = New System.Windows.Forms.ToolStripMenuItem()
         Me.BTT_CONTEXT_DELETE = New System.Windows.Forms.ToolStripMenuItem()
         Me.BTT_CONTEXT_FAV = New System.Windows.Forms.ToolStripMenuItem()
@@ -91,7 +97,6 @@ Partial Public Class MainFrame : Inherits System.Windows.Forms.Form
         Me.TRAY_CONTEXT = New System.Windows.Forms.ContextMenuStrip(Me.components)
         Me.BTT_TRAY_SHOW_HIDE = New System.Windows.Forms.ToolStripMenuItem()
         Me.BTT_TRAY_CLOSE = New System.Windows.Forms.ToolStripMenuItem()
-        Me.BTT_CONTEXT_DOWN_DATE_LIMIT = New System.Windows.Forms.ToolStripMenuItem()
         SEP_1 = New System.Windows.Forms.ToolStripSeparator()
         SEP_2 = New System.Windows.Forms.ToolStripSeparator()
         CONTEXT_SEP_1 = New System.Windows.Forms.ToolStripSeparator()
@@ -107,6 +112,7 @@ Partial Public Class MainFrame : Inherits System.Windows.Forms.Form
         MENU_VIEW_SEP_2 = New System.Windows.Forms.ToolStripSeparator()
         TRAY_SEP_1 = New System.Windows.Forms.ToolStripSeparator()
         MENU_VIEW_SEP_4 = New System.Windows.Forms.ToolStripSeparator()
+        MENU_DOWN_ALL_SEP_1 = New System.Windows.Forms.ToolStripSeparator()
         Me.Toolbar_TOP.SuspendLayout()
         Me.Toolbar_BOTTOM.SuspendLayout()
         Me.USER_CONTEXT.SuspendLayout()
@@ -188,6 +194,11 @@ Partial Public Class MainFrame : Inherits System.Windows.Forms.Form
         MENU_VIEW_SEP_4.Name = "MENU_VIEW_SEP_4"
         MENU_VIEW_SEP_4.Size = New System.Drawing.Size(141, 6)
         '
+        'MENU_DOWN_ALL_SEP_1
+        '
+        MENU_DOWN_ALL_SEP_1.Name = "MENU_DOWN_ALL_SEP_1"
+        MENU_DOWN_ALL_SEP_1.Size = New System.Drawing.Size(228, 6)
+        '
         'MENU_SETTINGS
         '
         Me.MENU_SETTINGS.AutoToolTip = False
@@ -207,7 +218,7 @@ Partial Public Class MainFrame : Inherits System.Windows.Forms.Form
         'Toolbar_TOP
         '
         Me.Toolbar_TOP.GripStyle = System.Windows.Forms.ToolStripGripStyle.Hidden
-        Me.Toolbar_TOP.Items.AddRange(New System.Windows.Forms.ToolStripItem() {Me.MENU_SETTINGS, SEP_1, Me.BTT_ADD_USER, Me.BTT_EDIT_USER, Me.BTT_DELETE_USER, Me.BTT_REFRESH, Me.BTT_SHOW_INFO, Me.BTT_CHANNELS, Me.BTT_DOWN_SAVED, SEP_2, Me.BTT_DOWN_SELECTED, Me.BTT_DOWN_ALL, Me.BTT_DOWN_VIDEO, Me.BTT_DOWN_STOP, SEP_3, Me.MENU_VIEW, SEP_4, Me.BTT_LOG, Me.BTT_VERSION_INFO, Me.BTT_DONATE})
+        Me.Toolbar_TOP.Items.AddRange(New System.Windows.Forms.ToolStripItem() {Me.MENU_SETTINGS, SEP_1, Me.BTT_ADD_USER, Me.BTT_EDIT_USER, Me.BTT_DELETE_USER, Me.BTT_REFRESH, Me.BTT_SHOW_INFO, Me.BTT_CHANNELS, Me.BTT_DOWN_SAVED, SEP_2, Me.BTT_DOWN_SELECTED, Me.MENU_DOWN_ALL, Me.BTT_DOWN_VIDEO, Me.BTT_DOWN_STOP, SEP_3, Me.MENU_VIEW, SEP_4, Me.BTT_LOG, Me.BTT_VERSION_INFO, Me.BTT_DONATE})
         Me.Toolbar_TOP.Location = New System.Drawing.Point(0, 0)
         Me.Toolbar_TOP.Name = "Toolbar_TOP"
         Me.Toolbar_TOP.Size = New System.Drawing.Size(934, 25)
@@ -285,14 +296,53 @@ Partial Public Class MainFrame : Inherits System.Windows.Forms.Form
         Me.BTT_DOWN_SELECTED.Text = "Download selected (F5)"
         Me.BTT_DOWN_SELECTED.ToolTipText = "Download selected user"
         '
+        'MENU_DOWN_ALL
+        '
+        Me.MENU_DOWN_ALL.AutoToolTip = False
+        Me.MENU_DOWN_ALL.DropDownItems.AddRange(New System.Windows.Forms.ToolStripItem() {Me.BTT_DOWN_ALL, Me.BTT_DOWN_SITE, MENU_DOWN_ALL_SEP_1, Me.BTT_DOWN_ALL_FULL, Me.BTT_DOWN_SITE_FULL})
+        Me.MENU_DOWN_ALL.Image = Global.SCrawler.My.Resources.Resources.StartPic_01_Green_16
+        Me.MENU_DOWN_ALL.ImageTransparentColor = System.Drawing.Color.Magenta
+        Me.MENU_DOWN_ALL.Name = "MENU_DOWN_ALL"
+        Me.MENU_DOWN_ALL.Size = New System.Drawing.Size(128, 22)
+        Me.MENU_DOWN_ALL.Text = "Download all (F6)"
+        '
         'BTT_DOWN_ALL
         '
+        Me.BTT_DOWN_ALL.AutoToolTip = True
         Me.BTT_DOWN_ALL.Image = Global.SCrawler.My.Resources.Resources.StartPic_01_Green_16
-        Me.BTT_DOWN_ALL.ImageTransparentColor = System.Drawing.Color.Magenta
         Me.BTT_DOWN_ALL.Name = "BTT_DOWN_ALL"
-        Me.BTT_DOWN_ALL.Size = New System.Drawing.Size(119, 22)
+        Me.BTT_DOWN_ALL.Size = New System.Drawing.Size(231, 22)
         Me.BTT_DOWN_ALL.Text = "Download all (F6)"
-        Me.BTT_DOWN_ALL.ToolTipText = "Download all users"
+        Me.BTT_DOWN_ALL.ToolTipText = "Download all users marked 'Ready for download' from all sites."
+        '
+        'BTT_DOWN_SITE
+        '
+        Me.BTT_DOWN_SITE.AutoToolTip = True
+        Me.BTT_DOWN_SITE.Image = Global.SCrawler.My.Resources.Resources.StartPic_01_Green_16
+        Me.BTT_DOWN_SITE.Name = "BTT_DOWN_SITE"
+        Me.BTT_DOWN_SITE.Size = New System.Drawing.Size(231, 22)
+        Me.BTT_DOWN_SITE.Text = "Download all site users"
+        Me.BTT_DOWN_SITE.ToolTipText = "Download all users marked 'Ready for download' from specific sites."
+        '
+        'BTT_DOWN_ALL_FULL
+        '
+        Me.BTT_DOWN_ALL_FULL.AutoToolTip = True
+        Me.BTT_DOWN_ALL_FULL.Image = Global.SCrawler.My.Resources.Resources.StartPic_01_Green_16
+        Me.BTT_DOWN_ALL_FULL.Name = "BTT_DOWN_ALL_FULL"
+        Me.BTT_DOWN_ALL_FULL.Size = New System.Drawing.Size(231, 22)
+        Me.BTT_DOWN_ALL_FULL.Text = "Download all [FULL]"
+        Me.BTT_DOWN_ALL_FULL.ToolTipText = "Download all users from all sites. The 'Ready for download' option will be ignore" &
+    "d."
+        '
+        'BTT_DOWN_SITE_FULL
+        '
+        Me.BTT_DOWN_SITE_FULL.AutoToolTip = True
+        Me.BTT_DOWN_SITE_FULL.Image = Global.SCrawler.My.Resources.Resources.StartPic_01_Green_16
+        Me.BTT_DOWN_SITE_FULL.Name = "BTT_DOWN_SITE_FULL"
+        Me.BTT_DOWN_SITE_FULL.Size = New System.Drawing.Size(231, 22)
+        Me.BTT_DOWN_SITE_FULL.Text = "Download all site users [FULL]"
+        Me.BTT_DOWN_SITE_FULL.ToolTipText = "Download all users from specific sites. The 'Ready for download' option will be i" &
+    "gnored."
         '
         'BTT_DOWN_VIDEO
         '
@@ -514,7 +564,7 @@ Partial Public Class MainFrame : Inherits System.Windows.Forms.Form
         '
         Me.USER_CONTEXT.Items.AddRange(New System.Windows.Forms.ToolStripItem() {Me.BTT_CONTEXT_DOWN, Me.BTT_CONTEXT_DOWN_LIMITED, Me.BTT_CONTEXT_DOWN_DATE_LIMIT, Me.BTT_CONTEXT_EDIT, Me.BTT_CONTEXT_DELETE, CONTEXT_SEP_1, Me.BTT_CONTEXT_FAV, Me.BTT_CONTEXT_TEMP, Me.BTT_CONTEXT_READY, Me.BTT_CONTEXT_GROUPS, Me.BTT_CONTEXT_ADD_TO_COL, Me.BTT_CONTEXT_COL_MERGE, Me.BTT_CONTEXT_CHANGE_FOLDER, CONTEXT_SEP_2, Me.BTT_CHANGE_IMAGE, CONTEXT_SEP_3, Me.BTT_CONTEXT_OPEN_PATH, CONTEXT_SEP_4, Me.BTT_CONTEXT_OPEN_SITE, CONTEXT_SEP_5, Me.BTT_CONTEXT_INFO})
         Me.USER_CONTEXT.Name = "USER_CONTEXT"
-        Me.USER_CONTEXT.Size = New System.Drawing.Size(222, 408)
+        Me.USER_CONTEXT.Size = New System.Drawing.Size(222, 386)
         '
         'BTT_CONTEXT_DOWN
         '
@@ -531,6 +581,13 @@ Partial Public Class MainFrame : Inherits System.Windows.Forms.Form
         Me.BTT_CONTEXT_DOWN_LIMITED.Size = New System.Drawing.Size(221, 22)
         Me.BTT_CONTEXT_DOWN_LIMITED.Text = "Download data limited"
         Me.BTT_CONTEXT_DOWN_LIMITED.ToolTipText = "Download top ... posts"
+        '
+        'BTT_CONTEXT_DOWN_DATE_LIMIT
+        '
+        Me.BTT_CONTEXT_DOWN_DATE_LIMIT.Image = Global.SCrawler.My.Resources.Resources.StartPic_01_Green_16
+        Me.BTT_CONTEXT_DOWN_DATE_LIMIT.Name = "BTT_CONTEXT_DOWN_DATE_LIMIT"
+        Me.BTT_CONTEXT_DOWN_DATE_LIMIT.Size = New System.Drawing.Size(221, 22)
+        Me.BTT_CONTEXT_DOWN_DATE_LIMIT.Text = "Download data to the date"
         '
         'BTT_CONTEXT_EDIT
         '
@@ -650,13 +707,6 @@ Partial Public Class MainFrame : Inherits System.Windows.Forms.Form
         Me.BTT_TRAY_CLOSE.Size = New System.Drawing.Size(133, 22)
         Me.BTT_TRAY_CLOSE.Text = "Close"
         '
-        'BTT_CONTEXT_DOWN_DATE_LIMIT
-        '
-        Me.BTT_CONTEXT_DOWN_DATE_LIMIT.Image = Global.SCrawler.My.Resources.Resources.StartPic_01_Green_16
-        Me.BTT_CONTEXT_DOWN_DATE_LIMIT.Name = "BTT_CONTEXT_DOWN_DATE_LIMIT"
-        Me.BTT_CONTEXT_DOWN_DATE_LIMIT.Size = New System.Drawing.Size(221, 22)
-        Me.BTT_CONTEXT_DOWN_DATE_LIMIT.Text = "Download data to the date"
-        '
         'MainFrame
         '
         Me.AutoScaleDimensions = New System.Drawing.SizeF(6.0!, 13.0!)
@@ -684,7 +734,7 @@ Partial Public Class MainFrame : Inherits System.Windows.Forms.Form
     Private WithEvents BTT_ADD_USER As ToolStripButton
     Private WithEvents BTT_DELETE_USER As ToolStripButton
     Private WithEvents BTT_DOWN_SELECTED As ToolStripButton
-    Private WithEvents BTT_DOWN_ALL As ToolStripButton
+    Private WithEvents MENU_DOWN_ALL As ToolStripDropDownButton
     Private WithEvents Toolbar_TOP As ToolStrip
     Private WithEvents Toolbar_BOTTOM As StatusStrip
     Private WithEvents PR_MAIN As ToolStripProgressBar
@@ -742,4 +792,8 @@ Partial Public Class MainFrame : Inherits System.Windows.Forms.Form
     Private WithEvents MENU_SETTINGS As ToolStripDropDownButton
     Private WithEvents BTT_PR_INFO As ToolStripStatusLabel
     Private WithEvents BTT_CONTEXT_DOWN_DATE_LIMIT As ToolStripMenuItem
+    Private WithEvents BTT_DOWN_ALL As ToolStripMenuItem
+    Private WithEvents BTT_DOWN_SITE As ToolStripMenuItem
+    Private WithEvents BTT_DOWN_ALL_FULL As ToolStripMenuItem
+    Private WithEvents BTT_DOWN_SITE_FULL As ToolStripMenuItem
 End Class